feat: Update Cart model to return total as float and add tests for total calculation

+++ app/Models/Cart.php
@@ -51,7 +51,7 @@ final class Cart extends Model
         return $this->products->isEmpty();
     }
 
-    public function total(): int
+    public function total(): float
     {
         $total = 0;
         foreach ($this->items as $cartItem) {
